Add compute_contract_address(), move some utility functions.
