+++ erdpy_core/address.py
@@ -1,7 +1,11 @@
 from typing import Tuple
 
-from erdpy_core import bech32, sharding
+from Cryptodome.Hash import keccak
+
+from erdpy_core import bech32
+from erdpy_core.constants import METACHAIN_ID
 from erdpy_core.errors import ErrBadAddress, ErrBadPubkeyLength
+from erdpy_core.interfaces import IAddress, INonce
 
 SC_HEX_PUBKEY_PREFIX = "0" * 16
 DEFAULT_HRP = "erd"
@@ -41,7 +45,7 @@ class Address():
         return self.hex().startswith(SC_HEX_PUBKEY_PREFIX)
 
     def get_shard(self) -> int:
-        shard = sharding.get_shard_of_pubkey(self.pubkey)
+        shard = get_shard_of_pubkey(self.pubkey)
         return shard
 
     def serialize(self) -> bytes:
@@ -103,3 +107,45 @@ def _decode_bech32(value: str) -> Tuple[str, bytes]:
         raise ErrBadAddress(value)
 
     return hrp, bytearray(decoded_bytes)
+
+
+def get_shard_of_pubkey(pubkey: bytes) -> int:
+    num_shards = 3
+    mask_high = int("11", 2)
+    mask_low = int("01", 2)
+
+    last_byte_of_pubkey = pubkey[31]
+
+    if _is_pubkey_of_metachain(pubkey):
+        return METACHAIN_ID
+
+    shard = last_byte_of_pubkey & mask_high
+    if shard > num_shards - 1:
+        shard = last_byte_of_pubkey & mask_low
+
+    return shard
+
+
+def _is_pubkey_of_metachain(pubkey: bytes) -> bool:
+    metachain_prefix = bytearray([0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0])
+    pubkey_prefix = pubkey[0:len(metachain_prefix)]
+    if pubkey_prefix == metachain_prefix:
+        return True
+
+    zero_address = bytearray(32)
+    if pubkey == zero_address:
+        return True
+
+    return False
+
+
+def compute_contract_address(deployer: IAddress, deployment_nonce: INonce, address_hrp: str) -> Address:
+    """
+    8 bytes of zero + 2 bytes for VM type + 20 bytes of hash(owner) + 2 bytes of shard(owner)
+    """
+    _, deployer_pubkey = _decode_bech32(deployer.bech32())
+    nonce_bytes = deployment_nonce.to_bytes(8, byteorder="little")
+    bytes_to_hash = deployer_pubkey + nonce_bytes
+    contract_pubkey = keccak.new(digest_bits=256).update(bytes_to_hash).digest()
+    contract_pubkey = bytes([0] * 8) + bytes([5, 0]) + contract_pubkey[10:30] + deployer_pubkey[30:]
+    return Address(contract_pubkey, address_hrp)
