+++ erdpy_core/address_test.py
@@ -2,7 +2,7 @@
 import pytest
 
 from erdpy_core.address import (Address, AddressConverter, AddressFactory,
-                                is_valid_bech32)
+                                compute_contract_address, is_valid_bech32)
 from erdpy_core.errors import ErrBadAddress, ErrBadPubkeyLength
 
 
@@ -56,6 +56,7 @@ def test_is_valid_bech32():
     assert not is_valid_bech32("foobar", "foo")
     assert not is_valid_bech32("erd1qyu5wthldzr8wx5c9ucg8kjagg0jfs53s8nr3zpz3hypefsdd8ssycr6th", "foo")
 
+
 def test_get_address_shard():
     address = Address.from_bech32("erd1qyu5wthldzr8wx5c9ucg8kjagg0jfs53s8nr3zpz3hypefsdd8ssycr6th")
     assert address.get_shard() == 1
@@ -65,3 +66,15 @@ def test_get_address_shard():
 
     address = Address.from_bech32("erd1k2s324ww2g0yj38qn2ch2jwctdy8mnfxep94q9arncc6xecg3xaq6mjse8")
     assert address.get_shard() == 2
+
+
+def test_compute_contract_address():
+    deployer = Address.from_bech32("erd1j0hxzs7dcyxw08c4k2nv9tfcaxmqy8rj59meq505w92064x0h40qcxh3ap")
+
+    contract_address = compute_contract_address(deployer, deployment_nonce=0, address_hrp="erd")
+    assert contract_address.pubkey.hex() == "00000000000000000500bb652200ed1f994200ab6699462cab4b1af7b11ebd5e"
+    assert contract_address.bech32() == "erd1qqqqqqqqqqqqqpgqhdjjyq8dr7v5yq9tv6v5vt9tfvd00vg7h40q6779zn"
+
+    contract_address = compute_contract_address(deployer, deployment_nonce=1, address_hrp="erd")
+    assert contract_address.pubkey.hex() == "000000000000000005006e4f90488e27342f9a46e1809452c85ee7186566bd5e"
+    assert contract_address.bech32() == "erd1qqqqqqqqqqqqqpgqde8eqjywyu6zlxjxuxqfg5kgtmn3setxh40qen8egy"
