+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "erdpy_core"
-version = "0.2.0"
+version = "0.3.0"
 authors = [
   { name="MultiversX" },
 ]
@@ -14,6 +14,9 @@ classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
+]
+dependencies = [
+  "pycryptodomex==3.16.0"
 ]
 
 [project.urls]
