+++ requirements.txt
@@ -1,0 +1,1 @@
+pycryptodomex==3.16.0
